requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pydantic==2.7.4
pydantic-settings==2.2.1

# LangGraph + OpenAI
langgraph==0.1.9
openai==1.43.0

# HubSpot
hubspot-api-client==9.0.0

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.30
alembic==1.13.1

# Background tasks
celery==5.4.0
redis==5.0.4

# HTTP
httpx==0.27.0
aiohttp==3.9.5

# Observability (optional but recommended)
langfuse==2.21.3